Move all enums to same .py file
