--- tibiapy/enums/gender.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class Gender(Enum):
-    FEMALE = "female"
-    MALE = "male"
