+++ tibiapy/enums/tibia_enums.py
@@ -1,6 +1,11 @@
 from enum import Enum
 
 
+class Gender(Enum):
+    FEMALE = "female"
+    MALE = "male"
+
+
 class HouseStatus(Enum):
     AUCTIONED = "auctioned"
     RENTED = "rented"
